rename: db->storage
